refactored s3 cloudfs

+++ src/scalarizr/storage2/cloudfs/s3.py
@@ -7,8 +7,8 @@ import os
 import sys
 
 from scalarizr.node import __node__
-from scalarizr.storage2.cloudfs.base import DriverError
-from scalarizr.storage2 import cloudfs
+from scalarizr.storage2.cloudfs.base import CloudFileSystem
+from scalarizr.storage2.cloudfs import cloudfs_types
 
 from boto.s3.key import Key
 from boto.exception import S3ResponseError
@@ -17,7 +17,7 @@ from boto.exception import S3ResponseError
 LOG = logging.getLogger(__name__)
 
 
-class S3FileSystem(object):
+class S3FileSystem(CloudFileSystem):
 	schema = 's3'
 	urlparse.uses_netloc.append(schema)  #?
 
@@ -25,44 +25,32 @@ class S3FileSystem(object):
 
 	_bucket = None
 
-
 	# TODO: change report frequency
 	def __init__(self, acl='aws-exec-read', report_frequency=11):
 		self.acl = acl
 		self.report_frequency = report_frequency
 
-	def parse_url(self, url):
-		o = urlparse.urlparse(url)
-		if o.scheme != self.schema:
-			raise DriverError('Wrong schema')
-		return o.hostname, o.path[1:]
+	def _parse_url(self, url):
+		bucket, key = super(S3FileSystem, self)._parse_url(url)
+		# while S3 web interface allows having bucket names with uppercase,
+		# boto throws exceptions when trying to manipulate them
+		return bucket.lower(), key
 
 	def ls(self, remote_path):
-		bucket_name, key_name = self.parse_url(remote_path)
+		bucket_name, key_name = self._parse_url(remote_path)
 
-		try:
-			connection = self._get_connection()
-			try:
-				if not self._bucket_check_cache(bucket_name):
-					self._bucket = connection.get_bucket(bucket_name, validate=False)
-			except S3ResponseError, e:
-				if e.code in ('NoSuchBucket',):
-					raise DriverError("S3 path '%s' not found" % remote_path)
-				raise
+		connection = self._get_connection()
 
-			files = [self._format_path(self._bucket.name, key.name)
-					 for key in self._bucket.list(prefix=key_name)]
-			return tuple(files)
-		except:
-			exc = sys.exc_info()
-			raise DriverError, exc[1], exc[2]
+		if not self._bucket_check_cache(bucket_name):
+			self._bucket = connection.get_bucket(bucket_name, validate=False)
 
-	def stat(self, path):
-		raise NotImplementedError()
+		files = [self._format_url(self._bucket.name, key.name)
+				 for key in self._bucket.list(prefix=key_name)]
+		return tuple(files)
 
 	def put(self, local_path, remote_path, report_to=None):
 		LOG.info("Uploading '%s' to S3 under '%s'", local_path, remote_path)
-		bucket_name, key_name = self.parse_url(remote_path)
+		bucket_name, key_name = self._parse_url(remote_path)
 		if key_name.endswith("/"):
 			key_name = os.path.join(key_name, os.path.basename(local_path))
 		LOG.debug("Uploading '%s'", key_name)
@@ -72,7 +60,7 @@ class S3FileSystem(object):
 
 			if not self._bucket_check_cache(bucket_name):
 				try:
-					bck = connection.get_bucket(bucket_name)
+					bck = connection.get_bucket(bucket_name, validate=False)  #?
 				except S3ResponseError, e:
 					if e.code == 'NoSuchBucket':
 						bck = connection.create_bucket(
@@ -94,7 +82,7 @@ class S3FileSystem(object):
 				key.set_contents_from_file(file_, policy=self.acl,
 					cb=report_to, num_cb=self.report_frequency)
 				LOG.debug("Finished uploading %s", os.path.basename(local_path))
-				return self._format_path(bucket_name, key_name)
+				return self._format_url(bucket_name, key_name)
 			finally:
 				if file_:
 					file_.close()
@@ -102,56 +90,37 @@ class S3FileSystem(object):
 		except:
 			exc = sys.exc_info()
 			LOG.debug('Caught error', exc_info=exc)
-			raise DriverError, exc[1], exc[2]
+			raise
 
 	def get(self, remote_path, local_path, report_to=None):
 		LOG.info('Downloading %s from S3 to %s', remote_path, local_path)
-		bucket_name, key_name = self.parse_url(remote_path)
+		bucket_name, key_name = self._parse_url(remote_path)
 		dest_path = os.path.join(local_path, os.path.basename(remote_path))
 
-		try:
-			connection = self._get_connection()
-			try:
-				if not self._bucket_check_cache(bucket_name):
-					self._bucket = connection.get_bucket(bucket_name, validate=False)
-				key = self._bucket.get_key(key_name)
-				if key is None:
-					raise DriverError("get_key returned None. No such key? %s" % key_name)
-			except S3ResponseError, e:
-				if e.code in ('NoSuchBucket', 'NoSuchKey'):
-					raise DriverError("S3 path '%s' not found" % remote_path)
-				raise
+		connection = self._get_connection()
 
-			LOG.debug("Actually downloading %s", os.path.basename(dest_path))
-			key.get_contents_to_filename(dest_path, cb=report_to,
-				num_cb=self.report_frequency)
-			LOG.debug("Finished downloading %s", os.path.basename(dest_path))
-			return dest_path
+		if not self._bucket_check_cache(bucket_name):
+			self._bucket = connection.get_bucket(bucket_name, validate=False)
+		key = self._bucket.get_key(key_name)
+		assert key, "No such key: %s" % key_name
 
-		except:
-			exc = sys.exc_info()
-			raise DriverError, exc[1], exc[2]
+		LOG.debug("Actually downloading %s", os.path.basename(dest_path))
+		key.get_contents_to_filename(dest_path, cb=report_to,
+			num_cb=self.report_frequency)
+		LOG.debug("Finished downloading %s", os.path.basename(dest_path))
+		return dest_path
 
 	def delete(self, remote_path):
 		LOG.info('Deleting %s from S3', remote_path)
-		bucket_name, key_name = self.parse_url(remote_path)
+		bucket_name, key_name = self._parse_url(remote_path)
 
-		try:
-			connection = self._get_connection()
-			try:
-				if not self._bucket_check_cache(bucket_name):
-					self._bucket = connection.get_bucket(bucket_name, validate=False)
-				key = self._bucket.get_key(key_name)
-				if key is None:
-					raise DriverError("get_key returned None. No such key? %s" % key_name)
-			except S3ResponseError, e:
-				if e.code in ('NoSuchBucket', 'NoSuchKey'):
-					raise DriverError("S3 path '%s' not found" % remote_path)
-				raise
-			return key.delete()
-		except:
-			exc = sys.exc_info()
-			raise DriverError, exc[1], exc[2]
+		connection = self._get_connection()
+
+		if not self._bucket_check_cache(bucket_name):
+			self._bucket = connection.get_bucket(bucket_name, validate=False)
+		key = self._bucket.get_key(key_name)
+
+		return key.delete() if key else None
 
 	def _bucket_check_cache(self, bucket):
 		if self._bucket and self._bucket.name != bucket:
@@ -170,8 +139,5 @@ class S3FileSystem(object):
 		else:
 			return region
 
-	def _format_path(self, bucket, key):
-		return '%s://%s/%s' % (self.schema, bucket, key)
 
-
-cloudfs.cloudfs_types["s3"] = S3FileSystem
+cloudfs_types["s3"] = S3FileSystem
